Create Maximum Sum of K Elements
You are given an array 
A
A containing 
N
N elements and an integer 
K
K. You have to find the subarray with the maximum sum among all the K-sized sub-arrays and output this maximum sum.

This is a perfect example of a question that can be efficiently solved using the sliding window approach.
Normally, we would calculate the sum of all K-sized sub-arrays which would have the time complexity of 
O
(
N
)
O(N).
But, using the Sliding Window approach we can assume the K-sized subarray to be a window. Now we just need to calculate the sum of the first window as the sum of the next window can be derived from the previous sum by subtracting the first element and adding the next element.

For example, let's take an array 
A
=
[
1
,
2
,
3
,
4
,
5
,
6
,
7
]
A=[1,2,3,4,5,6,7] and 
K
=
3
K=3.

Let's calculate the sum of the first window of three elements. 
1
+
2
+
3
=
6
1+2+3=6
Now the next window will have element 
2
,
3
,
4
2,3,4.
Compared to the previous window the first element 
(
1
)
(1) is removed and the next element 
(
4
)
(4) is added.
So, we can calculate the next window's sum by subtracting 
1
1 and adding 
4
4 to the sum of the previous window.
Thus, the sum of the next window will be 
6
−
1
+
4
=
9
6−1+4=9
Solve the above question using the Sliding Window approach

Input Format
The first line of the input contains a single integer 
N
N and 
K
K, denoting the length of the array 
A
A and the length of the sub-array.
The second line of the input contains 
N
N space-separated integers 
A
1
,
A
2
,
…
,
A
N
A 
1
​
 ,A 
2
​
 ,…,A 
N
​
  — denoting the array 
A
A.
Output Format
Output the maximum sum.
Constraints
2
≤
N
≤
1
0
5
2≤N≤10 
5
 
1
≤
A
i
≤
1
0
9
1≤A 
i
​
 ≤10 
9