Create Square Free Number
In the theory of numbers, square free numbers have a special place. A square free number is one that is not divisible by a perfect square (other than 
1
1). Thus 
72
72 is divisible by 
36
36 (a perfect square), and is not a square free number, but 
70
70 has factors 
1
,
2
,
5
,
7
,
10
,
14
,
35
1,2,5,7,10,14,35 and 
70
70. As none of these are perfect squares (other than 
1
1), 
70
70 is a square free number.

For some algorithms, it is important to find out the square free numbers that divide a number. Note that 
1
1 is not considered a square free number.

In this problem, you are asked to write a program to find the number of square free numbers that divide a given number.

Input Format
The only line of the input is a single integer 
N
N which is divisible by no prime number larger than 
19
19.