Create Left shift, Right shift
Left shift
Left-shifting a binary number involves moving its bits to the left by a specified number of positions.

In most of the languages, the left shift operator (<<) is used for this operation. The syntax is:

        result = original_value << num_positions
         11 << 2 = 1100 (Binary representation)
In binary representation, a left shift operation is equivalent to multiplying a number by 2 raised to the power of the shift amount.

Example of left shifting:
3 << 2 = 12 (Decimal representation)
3 * (2 2 ) = 12

Right shift
Right shifting a binary number involves moving its bits to the right by a specified number of positions.

In most of the languages, the right shift operator (>>) is used for this operation. The syntax is:

        result = original_value >> num_positions
        1100 >> 2 = 11 (Binary representation)
In binary representation, a right shift operation is equivalent to dividing a number by 2 raised to the power of the shift amount.

Example of right shifting:
12 >> 2 = 3 (Decimal representation)
12 / (2 2 ) = 3

Task
Input three integers N, L, and, R from user.
First, right-shift N by R positions, and then left-shift it by L positions
Print the final result.