Create The Mirror of Galadriel
The Mirror of Galadriel
We call a string S magical if every substring of S appears in Galadriel's Mirror (under lateral inversion). In other words, a magical string is a string where every substring has its reverse in the string.

Given a string S, determine if it is magical or not.

Input Format
The first line contains T, the number of test cases. The next T lines contain a string each.

Output Format
For each test case, output "YES" if the string is magical, and "NO" otherwise.